SushiBar/requirements.txt
asgiref==3.2.7
Babel==2.8.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
defusedxml==0.6.0
Django==3.0.5
django-allauth==0.41.0
django-ckeditor==5.9.0
django-cors-headers==3.6.0
django-crispy-forms==1.9.0
django-filter==2.4.0
django-js-asset==1.2.2
django-mathfilters==1.0.0
django-redis-cache==2.1.0
django-rq==2.1.0
django-suit==2.0a1
django-translation-manager==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
idna==2.9
markup==0.2
oauthlib==3.1.0
Pillow==7.1.2
polib==1.1.0
postgres==3.0.0
psycopg2-binary==2.8.5
psycopg2-pool==1.1
Pygments==2.6.1
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.3
redis==3.5.3
requests==2.23.0
requests-oauthlib==1.3.0
rq==1.5.2
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
